Número que a questão um recebia foi alterado pra float

+++ Listas/Lista 7/q1.py
@@ -4,7 +4,7 @@
 
 
 
-numero = int(input("Entre com o numero: "))
+numero = float(input("Entre com o numero: "))
 while numero != 999:
     triplo = numero * 3
     print("Triplo do números: {}".format(triplo))
